feat(flow): ALGO-FL-010 st_mincut_value — max-flow / min-cut duality wrapper, full 9-step SOP
`st_mincut_value` is a thin wrapper over `max_flow_value` justified by
the Ford-Fulkerson max-flow / min-cut theorem (1956). Mirrors
`igraph_st_mincut_value` at `references/igraph/src/flow/flow.c:1127`,
which is itself a 5-line redirect; we follow the same pattern and let
all input validation flow through the delegate so the error contract
is identical.

- 10 unit tests + 1 proptest invariant (`mincut_equals_maxflow`, 256
  cases) re-assert duality at the value level on random Graph fixtures;
  combined with FL-002's existing Edmonds-Karp cross-validation, the
  cut value inherits independent reference-checking transitively.
- 5 three-source conformance fixtures (c:2 / py:2 / r:1):
  - C 6-vertex directed from tests/unit/igraph_st_mincut_value.c (cut=7)
  - C 4-vertex undirected unit (cut=2 by parallel-path bottleneck)
  - py mirror of tests/test_flow.py:MinCutTests.testMinCutValue
  - R via min_cut(g, source, target, capacity), which dispatches to
    st_mincut_value_impl when both endpoints are given (R/flow.R:386).
- Bench `bench_st_mincut.rs` reuses the FL-002 fixtures so wrapper
  overhead is directly observable; back-to-back textbook measurements
  show max_flow_value=1.61us vs st_mincut_value=1.57us — zero
  measurable overhead. Snapshot in tracking/perf/ALGO-FL-010.json.
- Demo: `cargo run --example st_mincut_demo` (CLRS textbook + parallel
  paths + isolated endpoints + undirected igraph_maxflow.c reference).

Also cleans up pre-existing `--all-features` clippy lints in
`max_flow.rs:569,575,604,610` (`unnecessary_cast`, `cast_lossless`)
that CI was silently masking because CI's clippy invocation is
`--all-targets` without `--all-features`. The flow module is now
both CI-clean and all-features-clean.

+++ src/algorithms/flow/max_flow.rs
@@ -566,13 +566,13 @@ mod proptest_tests {
                 rng_state
             };
             let mut g = Graph::new(n, true).unwrap();
-            let edge_count = (next() as u32 % (n * 3 + 1)) as u32;
+            let edge_count = next() as u32 % (n * 3 + 1);
             let mut caps = Vec::with_capacity(edge_count as usize);
             for _ in 0..edge_count {
                 let u = (next() as u32) % n;
                 let v = (next() as u32) % n;
                 g.add_edge(u, v).unwrap();
-                let c = ((next() as u32 % 16) + 1) as f64;
+                let c = f64::from((next() as u32 % 16) + 1);
                 caps.push(c);
             }
             let source = 0;
@@ -601,13 +601,13 @@ mod proptest_tests {
                 rng_state
             };
             let mut g = Graph::with_vertices(n);
-            let edge_count = (next() as u32 % (n * 3 + 1)) as u32;
+            let edge_count = next() as u32 % (n * 3 + 1);
             let mut caps = Vec::with_capacity(edge_count as usize);
             for _ in 0..edge_count {
                 let u = (next() as u32) % n;
                 let v = (next() as u32) % n;
                 g.add_edge(u, v).unwrap();
-                let c = ((next() as u32 % 16) + 1) as f64;
+                let c = f64::from((next() as u32 % 16) + 1);
                 caps.push(c);
             }
             let source = 0;
